+++ integration-tests/src/kusama_test_net.rs
@@ -1,5 +1,5 @@
 #![cfg(test)]
-pub use basilisk_runtime::AccountId;
+pub use basilisk_runtime::{AccountId, VestingPalletId};
 use pallet_transaction_multi_payment::Price;
 use primitives::Balance;
 
@@ -180,6 +180,7 @@ pub fn basilisk_ext() -> sp_io::TestExternalities {
 			(AccountId::from(BOB), 1000 * UNITS),
 			(AccountId::from(CHARLIE), 1000 * UNITS),
 			(AccountId::from(DAVE), 1000 * UNITS),
+			(vesting_account(), 1_000_000 * UNITS)
 		],
 	}
 	.assimilate_storage(&mut t)
@@ -248,4 +249,8 @@ fn last_basilisk_events(n: usize) -> Vec<basilisk_runtime::Event> {
 
 pub fn expect_basilisk_events(e: Vec<basilisk_runtime::Event>) {
 	assert_eq!(last_basilisk_events(e.len()), e);
+}
+
+pub fn vesting_account() -> AccountId {
+	VestingPalletId::get().into_account()
 }
