+++ integration-tests/src/lib.rs
@@ -3,4 +3,5 @@ mod fees;
 mod kusama_test_net;
 mod nft_marketplace;
 mod non_native_fee;
+mod vesting;
 mod xyk;
